+++ wfmath/polygon.h
@@ -78,7 +78,7 @@ class Polygon<2>
 
   // Descriptive characteristics
 
-  int numCorners() const {return m_points.size();}
+  size_t numCorners() const {return m_points.size();}
   Point<2> getCorner(int i) const {return m_points[i];}
   Point<2> getCenter() const {return Barycenter(m_points);}
 
@@ -329,7 +329,7 @@ public:
 
   // Descriptive characteristics
 
-  int numCorners() const {return m_poly.numCorners();}
+  size_t numCorners() const {return m_poly.numCorners();}
   Point<dim> getCorner(int i) const {return m_orient.convert(m_poly[i]);}
   Point<dim> getCenter() const {return m_orient.convert(m_poly.getCenter());}
